ENYO-875: moon.VideoPlayer is not stop after pressing 'stop' button on remote control
Issue:
This sendFeedback is looking at slider.value to update the time on feedback. But, the jumpToStart call is set video.currentTime as 0 but not slider.value.
So, the current time on feedback popup is not updated to 0.

Fix:
We can set slider.value as 0 before calling jumpToStart.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com