ENYO-1030: Moving moon.History intialization out of create()
The moon.History singleton was doing some initialization in its
create() method that depended on the presence of webOS.js. This
was working fine in unminified app source (and probably in the
container as well), but not in deployed app source, because the
singleton was being created before webOS.js could be loaded.

To address this issue, we move the initialization code out of
create() and into a separate init() method that we invoke in a
call to enyo.ready().

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)