ENYO-1338: Add spotlight: 'container' to moon.Scrollable
Moonstone's scroller has historically been a Spotlight container.

I'm not entirely sure we need this to be the case going forward,
but changing it would probably have a bunch of hard-to-detect
implications. For example, the scroll controls assume that they're
operating in a container and we were seeing an issue with not being
able to 5-way up from the up arrow into a panel header above.

For now, simply ensuring that we get the same behavior as before by
adding spotlight: 'container' to moon.Scrollable. We can revisit
later if we care...

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)