ENYO-1444: moon.Scrollable: spot nearest child
When focus enters a scroller, we want to focus the nearest child
of the scroller, not the previously focused child.

This depends on a new Spotlight feature, `spotlightRememberFocus`,
which is a property that can be added to any Spotlight container
(defaults to true to preserve historical behavior).

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)