ENYO-1338: Make moon.Scrollable a Spotlight container
We need moon.Scrollable to be a Spotlight container so that 5-way
moves stay inside the scroller by preference.

However, When focus enters a scroller, we want to focus the
nearest child of the scroller, not the previously focused child.

This depends on a new Spotlight feature, `spotlightRememberFocus`,
which is a property that can be added to any Spotlight container
(defaults to true to preserve historical behavior).

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)