ENYO-1461: Scrollable: improve unspot-when-scrolling logic
enyo/Scrollable has an experimental `suppressMouseEvents` feature
to improve scrolling performance.

moon/Scrollable extends this behavior by calling
`spotlight/unspot()` when events are suppressed. This is necessary
to ensure that focus highlighting doesn't "stick" to the currently
focused control as the control scrolls out from under the pointer
(since it would normally be a mouse event that causes the highlight
to disappear).

Previously, we were calling `unspot()` regardless of whether the
currently focused element was in fact in the scrollable region (and
therefore moving). This was causing the Moonstone paging control
to be unhighlighted after it was tapped and scrolling began. To
avoid this, we now only call `unspot()` if the currently focused
element is inside the scrollable region.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)