ENYO-1054: Accessibility: add text to speech accessibility support to moonstone widgets, part 1
Apply accessibilty on moonstone ExpandablePicker and VideoPlayer.

ExpandablePicker:  When spotlight focused on Expandablepicker, the dom
which is focused has aria-label including only the content of
ExpandablePicker. In this case, the currentValue should be added in
aria-label as well, so observing the content of currentValue and add it
using accessibilityHint.

VideoPlayer: set accessibilityLabel to each icon button according to UX
document.

https://jira2.lgsvl.com/browse/ENYO-1054
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>