ENYO-1664: Avoid scrolling on "default" focus events
Spotlight can focus things in response to pointer events, 5-way
key events, or explicit requests to focus a particular Control. It
also sometimes focuses a "default" Control -- for example, when an
app is initialized, when the pointer disappears, or when an app
window regains focus.

These "default" focus cases can cause problems for a
moonstone/Scrollable. Specifically, a Scrollable will sometimes
mysteriously scroll to a different position when the pointer
disappears or the window regains focus. This happens when
Spotlight focuses an element (usually a recently focused element)
that is not currently in view, since most Moonstone elements
request to be scrolled into view upon focus.

To handle cases like these, we need to discriminate "default" focus
events from other types and override the behavior so that we focus
something that's already in view instead. We do so using the newly
added `focusType` property on the `onSpotlightFocus` event.

For now, we focus on the specific case of moonstone/NewDataList
because it's the one reported in ENYO-1664 and the easier one to
handle, but we'll also need to do something similar for any
moonstone/Scrollable, and we've noted that with inline comments.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)