refactor moonstone/Drawers to support CSS animation
Animation logic was spread between Drawers and Drawer with reliance on
enyo/Drawer's implementation. Due to the need to coordinate the
animation of a drawer and the client area of Drawers, we've tightly
coupled the two controls and therefore made Drawer a private kind.

Issue: ENYO-898
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)