add LunaService to moonstone/History
For the near term, this makes enyo-webos a required dependency for
Moonstone. That will likely need to change in the long term. It does
not, however, limit Moonstone apps to only webOS as the calls to
LunaService are guarded by a platform check and skip non-webOS
platforms.

Issue: ENYO-1496
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)