ENYO-1891: Add Text to Speech Accessibility support to Integer Picker
Add Text to Speech Accessibility support to Integer Picker
IntegerPicker uses FlyweightRepeater and it has two client items which
have same id, so it is hard to apply 'aria-labelledby'.
I just set accessibilityLabel on IntegerPicer for this.value. The reason
why I use accessibilityLabel instead of 'aria-label' is to be able to
use accessibilityHint if it is necessary.
For changing value, we can't use 'aria-live' because the
FlyweightRepeater replaces dom node whenever value changed, so just call
readAlert api for the value.

https://jira2.lgsvl.com/browse/ENYO-1891
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>