ENYO-2032: Accessibility: popup position is wrong.
When the popup is shown on 5 way mode, popup position is wrong.
Root cause is that if we set focus on dom node which is in hidden area,
web engine makes it scroll into showing area.
In this case when it set focus on button in the popup which is in hidden
area, web engine scroll up to show the button and then showing popup
animate works. This makes wrong position.
To resolve this, set the focus after animate end.

https://jira2.lgsvl.com/browse/ENYO-2032
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>