ENYO-1895: Add Text to Speech Accessibility support to Slider
I tried to apply aria attributes on valueChanged, but there are some
issues that if value is changed before animator complete TV reads
the value multiple times.
Set aria-valuenow and aria-valuetext in animatorComplete.
Initialize aria attributes on rendered to get popupContent.
When it initialized we set aria-hidden true to avoid first reading
that if aria-valuenow is set screen reader reads the value, and then
set it false when it spotFocused.

https://jira2.lgsvl.com/browse/ENYO-1895
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>