ENYO-2223: Prevent unwanted DOM scroll in NewDataList
When accessibility is enabled, the DOM focus() method is called on
any control that receives Spotlight focus (to trigger screen
reading). Some browser engines (including WebKit and Chrome) try to
force DOM-focused elements into view by adjusting the native DOM
scrollTop property of containing elements. This causes problems for
us, because we generally mange scrolling ourselves, and usually not
using native scrollTop.

In this case, we need to prevent scrollTop from being set on the
list's container element. We use a mechanism recently added to
address similar issues elsewhere in Moonstone.

TODO: See if we can find a clean, performant way to defeat this
type of behavior in general, unless a Control specifically opts in
to allow it -- because we pretty much never want the engine to
manage scrollTop outside our control.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)