ENYO-2224: Improve 5-way scrolling in Scrollable / NewDataList
* When doing continuous / accelerated 5-way scrolling, scroll an
element at a time instead of a page at a time. This produces
smoother scrolling.

* When 5-way focusing an element inside a scroller and focus was
previously outside the scroller or on one of the scroll controls,
minimize the distance scrolled to bring the element into view. This
produces a less jarring effect.

These changes necessitated some refactoring of existing logic for
filtering onSpotlightFocus events. The new factoring is explained
with inline comments.

The changes also depend on a new feature in Spotlight. The
onSpotlightFocus and onSpotlightFocused events now indicate
which element was previously focused, which enables handlers to
discriminate on that basis as necessary.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)