ENYO-2273-Scroller abnormal behavior on holding paging controls
Issue: Scroll bounds are calculated before rendering the content.
'bottomBoundary' property is dependent on scroll bounds, which is used
for comparison onpaginate scrollTo, methods . It's not updated during or
before the hold, if we use only 5-way to move on paging controls.
Fix: In scroller 'onSpotlightContainerEnter'  is handled on scroller to
update the scrollerbounds in 5-way
Enyo-DCO-1.1-Signed-off-by: Rajyavardhan P <rajyavardhan.p@lge.com>