ENYO-2334: Bonus fix -- close button positioning
The moon-popup-close class is intended to override the positioning
of the moon-icon class, but the specificity of the two selectors
was equal. We make the selector for moon-popup-close more specific,
so that positioning is overridden regardless of CSS order.

This issue was discovered while verifying the fix for ENYO-2334.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)