ENYO-2152: screen reader does not read accessibilityHint of component that has aria-labelledby
On expandable control the header control can get spotlight focus,
we set the aria-label on the header if it is needed.
The cases we should consider are
 - accessibilityLabel set
    : read only <accessibilityLabel>
 - accessibilityHint set
    : read <content> + <current value> + <accessibilityHint>
 - accessibilityLabel and accessibilityHint set
    : <accessibilityLabel> + <accessibilityHint>

apply similarly on ImageItem and LabeledTextItem.

apply below rules on GridListImageItem
 - accessibilityLabel set
    : read only <accessibilityLabel>
 - accessibilityHint set
    : read <caption> + <subCaption> + <accessibilityHint>
 - accessibilityLabel and accessibilityHint set
    : <accessibilityLabel> + <accessibilityHint>

change setAttribute to setAriaAttribute to checkt accessibilityDisabled

https://jira2.lgsvl.com/browse/ENYO-2152
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>