use renderInto when creating the internal marquee control
Using render() on the host control caused TTS to reread the content
because the focused div was dropped and then readded. Instead, use
renderInto to keep the host node, drop its content, and render the
marquee control within it. The resulting DOM is the same but avoids
dropping and readding the host node.

Issue: ENYO-2176
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)