ENYO-2502: Lost focus when create DataList dynamically on LightPanel
Issue:
When datalist is created dynamically and there is no focusable item,
lightPanels is give focus to spotlightPlaceholder.
But, datalist is fails to give focus when didRender is called because
current focus is not datalist and datalist is not trying to transfer
focus to inner item.

Fix:
Add condition to see the current focus is on spotlightPlaceholder or
spotlightDummy.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi <kunmyon.choi@lge.com>