ENYO-2468: Accessibility: Screen reader does not read the unit of DataPicker and TimePicker.
Issue
It doesn't read unit(day, month, hour etc) on DatePicker and TimePicker

Cause
aria-label is overrided in IntegerPicker after it is set by units
in TimePicker and DatePicker.

Fix
On IntergerPicker we just handle only accessibilityHint and aria-valuenow
to keep the accessibilityLabel which is set by TimePicker and DatePicker.
To do this we set the accessibilityRole of IntegerPicker to spinbutton and
only set the hint when spotlight is focused and then remove hint during
value changed.

https://jira2.lgsvl.com/browse/ENYO-2468
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>