ENYO-2685: It doesn't read accessibilityLabel of expandable item on ExpandablePicker
Issue
When the ExpandablePicker is focused, it doesn't read accessibilityLabel of Item but previous value.

Cause
We don't bind accessibilityLabel of expandable item to the ExpandablePicker.
So when the item is focused it reads accessibilityLabel but if ExpandablePicker
is focused it just reads previous value not accessibilityLabel.

Fix
When the selected is changed, bind accessibilityLabel of expandable items
to the accessibilityLabel of header text.

https://jira2.lgsvl.com/browse/ENYO-2685
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>