prevent acting on onEnd events in Slider when animating
On repeated key events, an animation can be started before the onEnd
fires. The result can be an incorrect value set because the value on
the animator is mid-animation and not snapped to the increment value.
To avoid this, we'll skip handling the onEnd event when an animation is
in progress.

Issue: ENYO-2638
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)