ENYO-2552: Accessibility: When the value of slider is zero, the hint of the narrow doesn't contain the value.
Issue
Accessibility: When the value of slider is zero, the hint of the narrow doesn't contain the value.

Cause
At the initial time when value is 0, 'this.$.popupLabel.getContent()' is empty.
So it doesn't read popup conent on left/right button.

Fix
Add null check for this.$.popupLabel.getContent().

https://jira2.lgsvl.com/browse/ENYO-2552
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>