ENYO-2710: "Scroll to boundary" support for Moonstone scroller
The primary way to scroll in Moonstone when in 5-way mode is to
move focus from one item to another and allow the scroller to
continually scroll the focused item into view as you go.

This means that if the elements closest to a scroller boundary
happen to be unfocusable (either because they're disabled or
because they don't support focus at all), you sometimes can't
scroll all the way to the edge of the scroller.

In this change, we add logic to detect when the just-focused
element is the last focusable element in the direction we're
moving. If it is the last focusable element, then we go ahead and
scroll all the way to the scroller boundary, provided that the
element will still be in view.

A partial solution to the same problem previously went in for
ENYO-2277, but that solution waited until focus left the scroller
to scroll to the boundary. We leave an enhanced version of that
solution intact, since it is still required to handle the case
where we can't scroll all the way to the boundary at the time of
focus because the last focusable element is far enough from the
scroller boundary that it would no longer be visible.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)