fire onChange after throttle timeout if value has changed
When quickly changing the value of a slider, the final one/few change
events can be suppressed by throttleJob resulting in the last onChange
event received by the UI being out of sync with the true value of the
Slider.

To resolve, start a job in the throttleJob callback with the same
duration that fires a final onChange if the value has changed since the
last time onChange was fired.

Issue: ENYO-2758
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)