fix scrollToBoundary check when leaving the container from a page ctrl
Previously checked against the kindName which is not always
predictable >= 2.6. Further, it assumed a kindName defined in
moonstone/ScrollStrategy which introduced some unnecessary coupling
such that a change in the kind used for paging controls would require a
change in moonstone/Scroller.

Instead, we'll use the ScrollStrategy's isPageControl() method, if it
exists, to check if the last spotted control should be checked for a
boundary. Also moved that check out of the job as it's inexpensive and
avoids a scenario in which the last control may have changed before the
job fires.

Issue: ENYO-2812
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)