specify vertical-align for RadioItem
The default value, baseline, produced extra space below the baseline to
account for characters that would below. As a result, the new scroll to
boundary logic would animate a slight scroll when the last RadioItem in
a scroller was spotted. Since RadioItem isn't designed to be used
inline with text, we'll specify the vertical-align to middle so all
are aligned correctly without the space below the baseline.

Issue: ENYO-2814
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)