end hold gesture when spotlight moves increment buttons
The logic within spotlight to handle cancelling the hold is subverted
because Slider disables the jump button onSpolightKeyDown (before
spotlight registers the button as the last 5-way control) so
_onDisappear doesn't cancel the hold (again because it wasn't
registered as the last 5-way control).

A more thorough refactoring of Slider would be warranted but manually
cancelling the hold when not accelerating and when the spotlight has
changed jump buttons resolves the issue in the interim.

Issue: ENYO-2815
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)