guard addControl with _created instead of generated
If controls are added to the SimplePicker after it is created but
before it is rendered, the internal state is updated correctly. The
guard appears to be there to avoid unnecessary work which will be done
during create() so we'll use the internal state member, _created, to
guard instead of generated.

Issue: ENYO-2597
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)