ENYO-2783: Change `scrollToControl()` to `scrollToChild()`
To fix ENYO-2783, we are doing some refactoring of enyo/Scrollable.
This requires a one-line change in moonstone/Scrollable to
accommodate a method-name change. (Not strictly necessary, as we
are supporting the legacy method name as well, but would like to
use the new one going forward.)

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)