ENYO-1224: ContextualPopup: Nothing is Spottable outside of Spotlight Modal Off
The components outside the contextual popup were not spottable even after the spotlightModal property set to false for the contextual popup. The spotlightModal property doesn't use for show/hide the scrim which caused this issue.

It was the modal property used for show/hide the scrim, so that the modal and spotlightModal are binded together and the this.showHideScrim() function called on modelChanged will solve the issue.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi kunmyon.choi@lge.com