refresh ExpandableDataPicker's DataRepeater when the collection changes
* ExpandablePicker uses the onActivate event bubbled from Checkbox to
  update its state.
* DataRepeater creates its children on render.
* ExpandableListItemDrawer defaults to renderOnShow

If the client code relies on the data in the collection to drive the
initial selection, it will not do so on create because the repeater has
not rendered yet so the onActivate event doesn't fire. So, if the
collection for an ExpandableDataPicker changes and the list hasn't
rendered yet, force a refresh to trigger the onActivate event, update
the selected item, and display the correct text in currentValueText.

Issue: ENYO-3101
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)