ENYO-3260 : Apply panel's accessibility case to other component.
In Panel Component, tts is not cut off even button is focused.
This is happen when special condition is satisfied.
(Fist change roll to 'alert', next dom.focus() is called)
In this case, webkit set clear option to false when call dom.focus()'s tts.
So, previous tts message is not cut off.
In this commit, the condition is applied to Popup & ContexturePopup Component.