ENYO-3260 : Change read sequence in popup, contexturePopup
In current status, when popup(or contexturePopup) is shown, read <button> + <title> sequence.

But UX guideline in Major MR, <title> + <button> sequence is right.

When onSpotlightContainerEnter is fired, the button is not focused yet.

So, if change role to alert in this handler, <alert> + <focus> rule is applied and TTS call is not cut off.

Enyo-DCO-1.1-Signed-off-by: Changgi Lee <changgi.lee@lge.com>