ENYO-3390: Flip start/end compare position on rtl case in moon.input
Issue:
When user press left or right key in moon.Input, it compares the cursor
position to 0 or length of text and move spotlight focus to other
control to that direction. Problem happens on rtl character. It compares
start position to 0 but it should compare to length.

Fix:
When rtl character is written in moon.Input, flip the start and end
position of compare. So it can safely move focus to the right direction.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)