ENYO-3379: Set ExpandableListItem spotlight property as false
Issue:
We added Scroll to boundary feature on ScrollStrategy which is checking
5way limit with its first spottable parent. When spotlight container
control is used in scroller with spottable child inside, it is detect
there is no nearest neighbor from the child and try to jump scroll.

This behavior is okay with other case but it is not when child items are
small number like ExpandableListItem Header case. It is try to jump
scroll for every 5way key.

Fix:
We changed spotlight property as container while adding back kay
handling on ExpandablePicker. But, ExpandableListItemDrawer is already
a container control. So, revert spotlight property back to false.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)