ENYO-3449: Fix tooltip arrow position for rtl text case
Issue:
Tooltip arrow is drawn as a svg element. It is absolutly positioned.
But, when tooltip is set direction rtl, sgv element moves its position
to opposit end. But, some right arrow case is having right: 0 and not
affected by direction change.

Fix:
We apply left: 0 for left arrow case also, so that it will not affected
by direction change.

Enyo-DCO-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)