ENYO-3447: Define .moon-word-break and use it in Notification
It's better to have word-break option to display context in various locales.
At this moment, we only applied it to BodyText in Notification.
For the expandability of word-break style, we separate it as an individual
css prop to use it in other controls.

Enyo-DCO-1.1-Signed-off-by: Yeram Choi (yeram.choi@lge.com)