ENYO-3473: Notification: text is rendered unclearly
This PR is for recording more agressive solution of this issue.

If we allow change button numbers or button content while popup is
showing, we need to monitor the status of all the buttons.
To do this, we can mixin default handlers for showingChanged,
contentChanged on createComponent.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)