When an item is selected in ExpandableDataPickerSample, it falls into an infinite loop. So, application is stopped.
Modified activateChanged handler is called when this.active is equeal to `true`.
Changed tap event handler of SelectableItem to be called when both 'selected' and 'active' are equal to `true`.

ENYO-4657 Enyo: when an item is selected in ExpandableDataPicker, it falls into an infinite loop
Enyo-DCO-1.1-Signed-off-by: Sangwook Lee sangwook1203.lee@lge.com